+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/drizzle
 
+## 0.1.36
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.12.0
+
 ## 0.1.35
 
 ### Patch Changes